chore: perubahan penyusunan modal & component, buang modal yang x guna letak dalam folder unused
