+++ components/modals/delete.js
@@ -2,10 +2,11 @@ import { useState } from 'react';
 import axios from 'axios';
 import { RiCloseLine } from 'react-icons/ri';
 
-import { BusyButton, SubmitButton } from './buttons';
-import styles from '../styles/Modal.module.css';
+import { toast } from 'react-toastify';
+import { BusyButton, SubmitButton } from '../buttons';
+import styles from '../../styles/Modal.module.css';
 
-const Modal = ({ setOpenDeleteModal, entity, mutate }) => {
+const Modal = ({ setOpenDeleteModal, toggle, entity, mutate }) => {
   const [deletingData, setDeletingData] = useState(false);
 
   const handleSubmit = async (e) => {
@@ -14,62 +15,18 @@ const Modal = ({ setOpenDeleteModal, entity, mutate }) => {
     try {
       const res = await axios.post('/gpass/api/gpass', {
         query: 'delete',
-        payload: { statusPegawai: entity.statusPegawai, bil: entity.bil },
+        payload: { type: toggle, bil: entity.bil },
       });
+      toast.success('Data berjaya dihapus');
       console.log(res);
     } catch (err) {
+      toast.error('Data gagal dihapus');
       console.log(err);
     }
     setDeletingData(false);
     setOpenDeleteModal(false);
     mutate();
   };
-
-  function BusyButton() {
-    return (
-      <>
-        <button
-          type='button'
-          className='inline-flex items-center text-center justify-center px-4 py-2 bg-admin3 text-adminWhite rounded-md shadow-xl p-2 hover:bg-admin1 transition-all ease-in-out duration-150 cursor-not-allowed'
-          disabled=''
-        >
-          <svg
-            className='animate-spin -ml-1 mr-3 h-5 w-5 text-white'
-            xmlns='http://www.w3.org/2000/svg'
-            fill='none'
-            viewBox='0 0 24 24'
-          >
-            <circle
-              className='opacity-25'
-              cx='12'
-              cy='12'
-              r='10'
-              stroke='currentColor'
-              strokeWidth='4'
-            ></circle>
-            <path
-              className='opacity-75'
-              fill='currentColor'
-              d='M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z'
-            ></path>
-          </svg>
-          Menghapus Jejakmu...
-        </button>
-      </>
-    );
-  }
-
-  function SubmitButton() {
-    return (
-      <button
-        type='submit'
-        onClick={handleSubmit}
-        className='capitalize bg-admin3 text-adminWhite rounded-md shadow-xl p-2 hover:bg-admin1 transition-all'
-      >
-        HAPUS
-      </button>
-    );
-  }
 
   return (
     <>
@@ -78,7 +35,7 @@ const Modal = ({ setOpenDeleteModal, entity, mutate }) => {
         onClick={() => setOpenDeleteModal(false)}
       />
       <div className={styles.centered}>
-        <div className={styles.modal}>
+        <div className={styles.modalDelete}>
           <div className={styles.modalHeader}>
             <h5 className={styles.heading}>AWAS!</h5>
           </div>
@@ -88,20 +45,22 @@ const Modal = ({ setOpenDeleteModal, entity, mutate }) => {
           >
             <RiCloseLine style={{ marginBottom: '-3px' }} />
           </button>
-          <div className={styles.modalContent}>
-            Anda YAKIN untuk menghapus {entity.nama}?
-          </div>
-          <div className={styles.modalActions}>
-            <div className={styles.actionsContainer}>
-              {deletingData ? <BusyButton /> : <SubmitButton />}
-              <button
-                className={styles.cancelBtn}
-                onClick={() => setOpenDeleteModal(false)}
-              >
-                Tidak
-              </button>
+          <form onSubmit={handleSubmit}>
+            <div className={styles.modalContent}>
+              Anda YAKIN untuk menghapus {entity.nama}?
             </div>
-          </div>
+            <div className={styles.modalActions}>
+              <div className={styles.actionsContainer}>
+                {deletingData ? <BusyButton /> : <SubmitButton func='del' />}
+                <button
+                  className={styles.cancelBtn}
+                  onClick={() => setOpenDeleteModal(false)}
+                >
+                  Tidak
+                </button>
+              </div>
+            </div>
+          </form>
         </div>
       </div>
     </>
