+++ components/modals/excel.js
@@ -2,11 +2,11 @@ import { useState } from 'react';
 import { RiCloseLine } from 'react-icons/ri';
 import axios from 'axios';
 
-import { BusyButton, SubmitButton } from './buttons';
-import styles from '../styles/Modal.module.css';
+import { BusyButton, SubmitButton } from '../buttons';
+import styles from '../../styles/Modal.module.css';
 import { toast } from 'react-toastify';
 
-const AddJson = ({ toggle, setAddJson }) => {
+const AddExcel = ({ toggle, setAddExcel }) => {
   const [addingData, setAddingData] = useState(false);
   const [success, setSuccess] = useState(false);
   const [count, setCount] = useState(0);
@@ -17,28 +17,32 @@ const AddJson = ({ toggle, setAddJson }) => {
       toast('No going back now!');
       return;
     } else {
-      setAddJson(false);
+      setAddExcel(false);
     }
   };
 
   const handleSubmit = async (e) => {
     e.preventDefault();
     setAddingData(true);
-    if (e.target.jsonFile.value === '') {
+    if (e.target.excelFile.value === '') {
       toast.error('Sila pilih fail');
       return;
     }
-    if (e.target.jsonFile.files[0].type !== 'application/json') {
-      toast.error('Fail bukan JSON');
+    if (
+      e.target.excelFile.files[0].type !==
+        'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet' &&
+      e.target.excelFile.files[0].type !== 'application/vnd.ms-excel'
+    ) {
+      toast.error('Fail bukan xlsx atau xls');
       return;
     }
-    const jsonFile = e.target.jsonFile.files[0];
+    const excelFile = e.target.excelFile.files[0];
     const formData = new FormData();
-    formData.append('jsonFile', jsonFile);
+    formData.append('excelFile', excelFile);
     formData.append('toggle', toggle);
     formData.append('addmode', addMode);
     try {
-      const res = await axios.post('/gpass/api/processjson', formData, {
+      const res = await axios.post('/gpass/api/processxlsx', formData, {
         headers: {
           'Content-Type': 'multipart/form-data',
         },
@@ -60,23 +64,23 @@ const AddJson = ({ toggle, setAddJson }) => {
         <div className={styles.centered}>
           <div className={styles.modalAdd}>
             <div className={styles.modalHeader}>
-              <h5 className={styles.heading}>TAMBAH DATA MENGGUNAKAN JSON</h5>
+              <h5 className={styles.heading}>TAMBAH DATA MENGGUNAKAN XLSX</h5>
             </div>
-            <button className={styles.closeBtn} onClick={noWayBack}>
+            <span className={styles.closeBtn} onClick={noWayBack}>
               <RiCloseLine style={{ marginBottom: '-3px' }} />
-            </button>
+            </span>
             <div className={styles.modalContent}>
-              <div className='mx-auto w-1/2 h-1/2 justify-center items-center mt-2'>
+              <div className='mx-auto w-2/3 h-1/2 justify-center items-center mt-2'>
                 <input
                   className='appearance-none bg-transparent border-none w-full text-gray-700 mr-3 py-1 px-2 leading-tight focus:outline-none'
                   type='file'
-                  id='jsonFile'
+                  id='excelFile'
                 />
                 <p className='text-gray-500 text-xs italic mt-5'>
                   {addingData ? 'Sabar...' : ''}
                 </p>
                 <p className='text-red-500 text-xs italic mt-5'>
-                  {addingData ? 'Memproses...' : 'Pilih File JSON'}
+                  {addingData ? 'Memproses...' : 'Pilih File Excel'}
                 </p>
                 <input
                   type='checkbox'
@@ -118,4 +122,4 @@ const AddJson = ({ toggle, setAddJson }) => {
   );
 };
 
-export default AddJson;
+export default AddExcel;
